Correcting su validation and tests when su => false
The su config option works differently than other logrotate config
options and does not have a corresponding 'nosu' version.  This adds
proper validation checks, tests and erb handling to allow for this.

Without this, it would be possible to have only 'su' in the logrotate
rule if no owner or group was specified.


Former-commit-id: d13811066f1ecfa779ada58ce1529499ad1edb0c