Default to ensure => 'present' for the package
Prior to this, the default value for the `ensure` parameter of the
logrotate package was set to 'latest'. This meant that by simply
declaring a logrotate::rule, the package could also be updated. It is
generally not a good idea to update system packages unless being done in
a controlled envrionment.

To fix that issue, the default value is being changed to 'present'.
This way, the end-user can choose to update the logrotate package when
they want to.

This fixes issue #11