Remove UserOrGroup type, use String
The regex that matches a valid user or group is operating system and
distribution dependent. For instance, for older distributions the
contents of the useradd man page will often give a recommended regex,
rather than a required one.

Remove the UserOrGroup data type and use String instead to account for
horribly named users and groups.